% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banana_split.R
\name{banana_split}
\alias{banana_split}
\title{Tidy a string for splitting in half.}
\usage{
banana_split(x, sep = " ", pos = 1, join = " ")
}
\arguments{
\item{x}{A string containing characters with different separators.}

\item{sep}{Character element at which the resulting string will be split. The default is a single space. Letters are allowed and will be retained in the original character element. Not working if multiple letters occur as separators.}

\item{pos}{Integer giving the separator position if it occurs twice in a row. Not working if separator occurs more than twice in a row.}

\item{join}{Character element joining the two resulting strings. The default is a single space.}
}
\value{
A string of two character elements joined by character element of choice.
}
\description{
Cleans up a messy string so that it can be split at the separator of choice. The string may contain different separators. A two-element string is not ordered and returned as is.
}
\examples{
# a string with different separators
string1 <- "blue_mono green mixed"; string1
banana_split(string1, sep = " ")

# a string with a letter as separator that also occurs as part of the character elements
string2 <- "bluedmonodtypedgreendmixeddtype"; string2
banana_split(string2, sep = "d", join = "_")

# first character is separator, keep second occurrence in a row (as part of element)
string3 <- "blue_monoggreengmixed"; string3
banana_split(string3, sep = "g", pos=1, join = " ")

# second character is separator, keep first occurrence in a row (as part of element)
string4 <- "mono.yellowwmixed red"; string4
banana_split(string4, sep = "w", pos=2, join = " ")

# will not change anything in the string (already two elements present)
string5 <- "1 2"; string5
banana_split(string5, sep = "_") # separator is ignored.
}
