% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISTD.R
\name{ISTD}
\alias{ISTD}
\title{Standardize data based on internal standard}
\usage{
ISTD(data, remove_ISTD = F, show_original = T, save_as_textfile = F)
}
\arguments{
\item{data}{A data frame to be standardized.}

\item{remove_ISTD}{Logical. Should the column containing the ISTD be removed after standardization? \code{FALSE} by default.}

\item{show_original}{Logical. Should the original data be returned additionally? \code{TRUE} by default.}

\item{save_as_textfile}{Logical. Should the standardized output be saved in the working directory? \code{FALSE} by default.}
}
\value{
A list of tidy data frames of standardized data and, optionally, the original data. Additionally, the list returns information on what samples where ISTD was not detected.
}
\description{
This function standardizes a given data frame based on an internal standard (ISTD).
}
\examples{
\dontrun{
# generate example data
set.seed(123)
df <- data.frame(replicate(10, sample(1:10000, 100, T))) \%>\%
purrr::set_names(wakefield::name(10)) \%>\%
mutate(Sample = stringi::stri_rand_strings(100, 5),
       Treatment = sample(LETTERS[1:3], 100, T),
       compound_ISTD = sample(c(NA, 100:120), 100, T),
       .before = everything(.)) \%>\%
       tibble(); df

# standardize example data
standardised_output <- ISTD(data,
                            remove_ISTD = F,
                            show_original = T,
                            save_as_textfile = F)
dframe_ISTD$standardised_data
dframe_ISTD$original_data
dframe_ISTD$no_ISTD
}
}
