% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_GCMSsolution.R
\name{parse_GCMSsolution}
\alias{parse_GCMSsolution}
\title{Tidy batch processing output}
\usage{
parse_GCMSsolution(
  file,
  samples,
  compounds,
  results_file,
  ref_ions = 3,
  save_as_textfiles = F
)
}
\arguments{
\item{file}{A string containing the name of the output file after batch processing. If not renamed, the default file name suggested by the software will be "ASCIIData.txt".}

\item{samples}{An integer giving the number of samples analyzed.}

\item{compounds}{An integer giving the number of compounds scored (as shown in the compound table of the method file).}

\item{results_file}{A string giving the file name to save the tidy output.}

\item{ref_ions}{An integer giving the number of reference ions (m/z) to add (default is three, maximum is five).}

\item{save_as_textfiles}{Logical. Should the cleaned output be saved in the working directory? \code{FALSE} by default.}
}
\value{
A list of tidy output from the cleaned batch processing. The first element contains a tibble with sample names as rows and the compound areas as columns. The second element contains a tibble with additional information for each samples like base peak, peak height, retention index, and retention time. The third element contains a tibble with the summary of all detected compounds listed with their base peak, and the median of their retention index and retention time.
}
\description{
Creates a list of tidy output from batch processing based on post-run analysis in Shimadzu GSCMSsolution software v4.45. The function is a parser for output files from batch processing with the "Compound Quantitative Result" checkbox selected in the settings.
}
\examples{
# example output file
\dontrun{
tidy_output <- parse_GCMSsolution(system.file(package="SEmisc",
                                    "extdata/example_batchprocessing_GCMSsolution.txt"),
                                    samples = 3,
                                    compounds = 80,
                                    results_file = "tidy_output",
                                    ref_ions = 3,
                                    save_as_textfiles = F)
tidy_output[[1]] # tidy output in wide format
tidy_output[[2]] # tidy output in long format
tidy_output[[3]] # tidy compound summary}
}
